+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,13 +1,13 @@
-package kitchenpos.ordertable.application;
+package kitchenpos.tablegroup.application;
 
 import kitchenpos.ordertable.application.dto.CreateOrderTableIdDto;
-import kitchenpos.ordertable.application.dto.CreateTableGroupDto;
-import kitchenpos.ordertable.application.dto.TableGroupDto;
+import kitchenpos.tablegroup.application.dto.CreateTableGroupDto;
+import kitchenpos.tablegroup.application.dto.TableGroupDto;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.ordertable.domain.OrderValidator;
-import kitchenpos.ordertable.domain.TableGroup;
-import kitchenpos.ordertable.domain.TableGroupRepository;
+import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroupRepository;
 import kitchenpos.ordertable.exception.OrderTableException;
 import kitchenpos.ordertable.exception.TableGroupException;
 import org.springframework.stereotype.Service;
