+++ src/test/java/kitchenpos/ordertable/application/TableGroupServiceTest.java
@@ -2,17 +2,18 @@ package kitchenpos.ordertable.application;
 
 import kitchenpos.MockServiceTest;
 import kitchenpos.ordertable.application.dto.CreateOrderTableIdDto;
-import kitchenpos.ordertable.application.dto.CreateTableGroupDto;
+import kitchenpos.tablegroup.application.dto.CreateTableGroupDto;
 import kitchenpos.ordertable.application.dto.OrderTableDto;
-import kitchenpos.ordertable.application.dto.TableGroupDto;
+import kitchenpos.tablegroup.application.dto.TableGroupDto;
 import kitchenpos.ordertable.domain.GuestNumber;
 import kitchenpos.order.domain.Order;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.ordertable.exception.OrderTableException;
 import kitchenpos.ordertable.exception.TableGroupException;
 import kitchenpos.ordertable.domain.OrderTableRepository;
-import kitchenpos.ordertable.domain.TableGroupRepository;
+import kitchenpos.tablegroup.domain.TableGroupRepository;
+import kitchenpos.tablegroup.application.TableGroupService;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
